make benchmarks results Windows-friendly
fixes #530
