+++ .github/workflows/ci.yml
@@ -8,6 +8,10 @@ on:
     # every day at 10pm CST (4 UTC)
     - cron: '0 4 * * *'
 
+env:
+  macosGradleArgs: '-Dorg.gradle.jvmargs=-Xmx10g -Dfile.encoding=UTF-8 -XX:+UseParallelGC'
+  ubuntuGradleArgs: '-Dorg.gradle.jvmargs=-Xmx5g -Dfile.encoding=UTF-8 -XX:+UseParallelGC'
+  windowsGradleArgs: '-Dorg.gradle.jvmargs=-Xmx4g -Dfile.encoding=UTF-8 -XX:+UseParallelGC'
 
 jobs:
 
@@ -266,7 +270,9 @@ jobs:
       - name: all tests
         uses: gradle/gradle-build-action@v2
         with:
-          arguments: testJvm
+          arguments: |
+            testJvm
+            "${{ env.ubuntuGradleArgs }}"
           cache-read-only: false
 
   instrumentation-tests:
@@ -310,3 +316,27 @@ jobs:
         with:
           name: insrumentation-test-results
           path: ./**/build/reports/androidTests/connected/**
+
+  tests-windows:
+
+    runs-on: windows-latest
+
+    steps:
+      - uses: actions/checkout@v3
+        with:
+          token: ${{ secrets.GITHUB_TOKEN }}
+          fetch-depth: 0
+
+      - name: Set up JDK
+        uses: actions/setup-java@v3
+        with:
+          distribution: 'zulu'
+          java-version: '11'
+
+      - name: all tests
+        uses: gradle/gradle-build-action@v2
+        with:
+          arguments: |
+            testJvm
+            "${{ env.windowsGradleArgs }}"
+          cache-read-only: false
